Merge git://git.kernel.org/pub/scm/linux/kernel/git/davem/net
BPF alignment tests got a conflict because the registers
are output as Rn_w instead of just Rn in net-next, and
in net a fixup for a testcase prohibits logical operations
on pointers before using them.

Also, we should attempt to patch BPF call args if JIT always on is
enabled.  Instead, if we fail to JIT the subprogs we should pass
an error back up and fail immediately.

[huexxx@gmail.com: only mods in verifier.c, include also get_callee_stack_depth()]
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 420bd19cbfcbdeebc8ac7f146b2cb1beb2e9a5b0)